Move auth routes to blueprint and out of server Removed commented out code. Fix login.view config
